# http-handler/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/profile_picture/{id}": {
            "post": {
                "description": "Uploads a profile picture to an S3 bucket",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "upload"
                ],
                "summary": "Upload a profile picture to S3",
                "parameters": [
                    {
                        "type": "file",
                        "description": "Image file to upload",
                        "name": "image",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "400": {
                        "description": "Bad request or file not found",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    }
                }
            }
        },
        "/ranking/court/{id}": {
            "get": {
                "description": "Retourne la liste des utilisateurs et leur ELO pour un court donné, triée par ELO croissant",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ranking"
                ],
                "summary": "Classement ELO par terrain",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Identifiant du court",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.CourtRankingResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "ID manquant",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    },
                    "401": {
                        "description": "Utilisateur non autorisé",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    },
                    "500": {
                        "description": "Erreur serveur / base",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    }
                }
            }
        },
        "/ranking/user/{userId}": {
            "get": {
                "description": "Retourne uniquement la liste des fields associés à un utilisateur (ex: terrains classés/évalués)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Liste des terrains (fields) d’un utilisateur",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Identifiant de l'utilisateur",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Field"
                            }
                        }
                    },
                    "400": {
                        "description": "userId manquant",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    },
                    "401": {
                        "description": "Utilisateur non autorisé",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    },
                    "500": {
                        "description": "Erreur serveur / base",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    }
                }
            }
        },
...
        "/users/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve param information, including profile picture and preferences",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Get a param by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.UserResponse"
                        }
                    },
                    "400": {
                        "description": "Missing ID in URL params",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    },
                    "404": {
                        "description": "User not found",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete a user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Delete a user by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Missing ID in URL params",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    },
                    "403": {
                        "description": "Incorrect rights",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Patch a user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Patch a user by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "User fields to update",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UserPatchRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Missing ID in URL params",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    },
                    "403": {
                        "description": "Incorrect rights",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "models.CourtRankingResponse": {
            "type": "object",
            "properties": {
                "elo": {
                    "type": "integer"
                },
                "userId": {
                    "type": "string"
                }
            }
        },
...
        "models.Error": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "models.Field": {
            "type": "object",
            "properties": {
                "elo": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "ranking": {
                    "type": "integer"
                }
            }
        },
...
        "models.Sport": {
            "type": "string",
            "enum": [
                "basket",
                "foot",
                "ping-pong"
            ],
            "x-enum-varnames": [
                "Basket",
                "Foot",
                "PingPong"
            ]
        },
...
        "models.UserPatchRequest": {
            "type": "object",
            "properties": {
                "bio": {
                    "description": "@nullable",
                    "type": "string"
                },
                "currentFieldId": {
                    "description": "@nullable",
                    "type": "string"
                },
                "email": {
                    "description": "@nullable",
                    "type": "string"
                },
                "username": {
                    "description": "@nullable",
                    "type": "string"
                }
            }
        },
        "models.UserResponse": {
            "type": "object",
            "properties": {
                "bio": {
                    "description": "@nullable",
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "currentFieldId": {
                    "type": "string"
                },
                "favoriteCity": {
                    "description": "@nullable",
                    "type": "string"
                },
                "favoriteField": {
                    "description": "@nullable",
                    "type": "string"
                },
                "favoriteSport": {
                    "description": "@nullable",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.Sport"
                        }
                    ]
                },
                "fields": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Field"
                    }
                },
                "nbMatches": {
                    "type": "integer"
                },
                "profilePicture": {
                    "description": "@nullable",
                    "type": "string"
                },
                "sports": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Sport"
                    }
                },
                "username": {
                    "type": "string"
                },
                "visitedFields": {
                    "type": "integer"
                },
                "winrate": {
                    "description": "@nullable",
                    "type": "integer"
                }
            }
        }
    }
}